Added db driver data to Query
Unfortunately, Query will have to be aware of the database for which the
query is being constructed. For example, pgsql does not support the
lastInsertID PDO function so the query will have to be modified to return
the new ID on INSERT queries.

This commit adds driver name to Query, and driver parsing to DB.